requirement.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django @ file:///tmp/build/80754af9/django_1609784082548/work
django-ckeditor==6.0.0
django-js-asset==1.2.2
djangorestframework==3.13.1
html5lib==1.1
idna==2.10
lxml==4.6.3
Pillow==8.1.0
psycopg2 @ file:///C:/ci/psycopg2_1608129576564/work
pytz @ file:///tmp/build/80754af9/pytz_1608922264688/work
requests==2.25.1
six==1.15.0
soupsieve==2.2
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
urllib3==1.26.3
webencodings==0.5.1
wincertstore==0.2
